requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
discord.py==2.5.2
frozenlist==1.6.0
h11==0.16.0
idna==3.10
multidict==6.4.3
outcome==1.3.0.post0
packaging==25.0
plyer==2.1.0
propcache==0.3.1
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
schedule==1.2.2
selenium==4.32.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.0
